requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
paramiko==3.3.1
pyyaml==6.0.1
requests==2.31.0
